update docstrings and tidy code
